feat(kbar): add "Create field" contextual action on records table
Add a command-palette contextual action that opens the custom-field
form inside the palette, targeted at the records table's entity.

To host the form un-modaled, split CustomFieldDialog into a shell-less
CustomFieldForm core (all field logic) plus a thin modal wrapper that
preserves the existing public API — the 6 existing modal callers are
untouched. The palette hosts the same core via a new create-field page,
mirroring how Create record uses EntityInstanceForm/CreatePage.

RELATIONSHIP springs the palette page to xxl via a reactive DialogNavPage
size (driven by the same onWideChange the modal uses) without remounting
the page, so form state is preserved.
